refactor: extract CoS task prompt builders (#4898)
